requirements.txt
Flask==2.3.3
Flask-Login==0.6.3
Werkzeug==2.3.8
gunicorn==20.1.0
psycopg2-binary==2.9.7
SQLAlchemy==2.0.20
python-dotenv==1.0.0
